data_collection_pipeline/requirements.txt
# QuantVerse Data Collection Pipeline Requirements

# Core dependencies
asyncio
asyncpg>=0.28.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Financial data APIs
yfinance>=0.2.0

# Text processing (optional - for enhanced sentiment analysis)
textblob>=0.17.0

# XML processing (built-in)
# xml.etree.ElementTree - built-in

# JSON processing (built-in)
# json - built-in

# Date/time processing (built-in)
# datetime - built-in

# HTTP processing (built-in with aiohttp)
# urllib.parse - built-in

# Email utilities (built-in)
# email.utils - built-in

# Hashing (built-in)
# hashlib - built-in

# Regular expressions (built-in)
# re - built-in

# Type hints (built-in in Python 3.9+)
# typing - built-in

# Async support (built-in)
# asyncio - built-in

# Signal handling (built-in)
# signal - built-in

# System utilities (built-in)
# sys - built-in

# Path utilities (built-in)
# pathlib - built-in

# Logging (built-in)
# logging - built-in

# Random (built-in)
# random - built-in

# Time utilities (built-in)
# time - built-in
